Absolute Value of the Relative Contribution of IPCs to the Interaction

ZA.AMMI computes the Absolute Value of the Relative Contribution of IPCs to the Interaction (\(\textrm{Z}_{\textrm{a}}\)) (Zali et al. 2012) considering all significant interaction principal components (IPCs) in the AMMI model. Using \(\textrm{Z}_{\textrm{a}}\), the Simultaneous Selection Index for Yield and Stability (SSI) is also calculated according to the argument ssi.method.

Usage

ZA.AMMI(model, n, alpha = 0.05, ssi.method = c("farshadfar", "rao"), a = 1)

Arguments

model

The AMMI model (An object of class AMMI generated by AMMI).

n

The number of principal components to be considered for computation. The default value is the number of significant IPCs.

alpha

Type I error probability (Significance level) to be considered to identify the number of significant IPCs.

ssi.method

The method for the computation of simultaneous selection index. Either "farshadfar" or "rao" (See SSI).

a

The ratio of the weights given to the stability components for computation of SSI when method = "rao" (See SSI).

Value

A data frame with the following columns:

Za

The Za values.

SSI

The computed values of simultaneous selection index for yield and stability.

rZa

The ranks of Za values.

rY

The ranks of the mean yield of genotypes.

means

The mean yield of the genotypes.

The names of the genotypes are indicated as the row names of the data frame.

Details

The Absolute Value of the Relative Contribution of IPCs to the Interaction (\(Za\)) (Zali et al. 2012) is computed as follows:

\[Za = \sum_{i=1}^{N'}\left | \theta_{n}\gamma_{in} \right |\]

Where, \(N'\) is the number of significant IPCAs (number of IPC that were retained in the AMMI model via F tests); \(\gamma_{in}\) is the eigenvector value for \(i\)th genotype; and \(\theta_{n}\) is the percentage sum of squares explained by the \(n\)th principal component interaction effect..

References

Zali H, Farshadfar E, Sabaghpour SH, Karimizadeh R (2012). “Evaluation of genotype × environment interaction in chickpea using measures of stability from AMMI model.” Annals of Biological Research, 3(7), 3126–3136.

See also

Examples

library(agricolae)
data(plrv)

# AMMI model
model <- with(plrv, AMMI(Locality, Genotype, Rep, Yield, console = FALSE))

# ANOVA
model$ANOVA
#> Analysis of Variance Table
#> 
#> Response: Y
#>            Df Sum Sq Mean Sq  F value    Pr(>F)    
#> ENV         5 122284 24456.9 257.0382  9.08e-12 ***
#> REP(ENV)   12   1142    95.1   2.5694  0.002889 ** 
#> GEN        27  17533   649.4  17.5359 < 2.2e-16 ***
#> ENV:GEN   135  23762   176.0   4.7531 < 2.2e-16 ***
#> Residuals 324  11998    37.0                       
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

# IPC F test
model$analysis
#>     percent  acum Df     Sum.Sq   Mean.Sq F.value   Pr.F
#> PC1    56.3  56.3 31 13368.5954 431.24501   11.65 0.0000
#> PC2    27.1  83.3 29  6427.5799 221.64069    5.99 0.0000
#> PC3     9.4  92.7 27  2241.9398  83.03481    2.24 0.0005
#> PC4     4.3  97.1 25  1027.5785  41.10314    1.11 0.3286
#> PC5     2.9 100.0 23   696.1012  30.26527    0.82 0.7059

# Mean yield and IPC scores
model$biplot
#>         type    Yield         PC1          PC2         PC3         PC4
#> 102.18   GEN 26.31947 -1.50828851  1.258765244 -0.19220309  0.48738861
#> 104.22   GEN 31.28887  0.32517729 -1.297024517 -0.63695749 -0.44159957
#> 121.31   GEN 30.10174  0.95604605  1.143461054 -1.28777348  2.22246913
#> 141.28   GEN 39.75624  2.11153737  0.817810467  1.45527701  0.25257620
#> 157.26   GEN 36.95181  1.05139017  2.461179974 -1.97208942 -1.96538800
#> 163.9    GEN 21.41747 -2.12407441 -0.284381234 -0.21791137 -0.50743629
#> 221.19   GEN 22.98480 -0.84981828  0.347983673 -0.82400783 -0.11451944
#> 233.11   GEN 28.66655  0.07554203 -1.046497338  1.04040485  0.22868362
#> 235.6    GEN 38.63477  1.20102029 -2.816581184  0.80975361  1.02013062
#> 241.2    GEN 26.34039 -0.79948495  0.220768053 -0.98538801  0.30004421
#> 255.7    GEN 30.58975 -1.49543817 -1.186549449  0.92552519 -0.32009239
#> 314.12   GEN 28.17335  1.39335380 -0.332786322 -0.73226877  0.05987348
#> 317.6    GEN 35.32583  1.05170769  0.002555823 -0.81561907  0.58180433
#> 319.20   GEN 38.75767  3.08338144  1.995946966  0.87971668 -1.11908943
#> 320.16   GEN 26.34808 -1.55737097  0.732314249 -0.41432567  1.32097009
#> 342.15   GEN 26.01336 -1.35880873 -0.741980068  0.87480105 -1.12013125
#> 346.2    GEN 23.84175 -2.48453928 -0.397045286  1.07091711 -0.90974484
#> 351.26   GEN 36.11581  1.22670345  1.537183139  1.79835728 -0.03516368
#> 364.21   GEN 34.05974  0.27328985 -0.447941156  0.03139543  0.77920500
#> 402.7    GEN 27.47748 -0.12907269 -0.080086669  0.01934016 -0.36085862
#> 405.2    GEN 28.98663 -1.90936369  0.309047963  0.57682642  0.51163370
#> 406.12   GEN 32.68323  0.90781100 -1.733433781 -0.24223050 -0.38596144
#> 427.7    GEN 36.19020  0.42791957 -0.723190970 -0.85381724 -0.53089914
#> 450.3    GEN 36.19602  1.38026196  1.279525147  0.16025163  0.61270137
#> 506.2    GEN 33.26623 -0.33054261 -0.302588536 -1.58471588 -0.04659416
#> Canchan  GEN 27.00126  1.47802905  0.380553178  1.67423900  0.07718375
#> Desiree  GEN 16.15569 -3.64968796  1.720025405  0.43761089  0.04648011
#> Unica    GEN 39.10400  1.25331924 -2.817033826 -0.99510845 -0.64366599
#> Ayac     ENV 23.70254 -2.29611851  0.966037760  1.95959116  2.75548057
#> Hyo-02   ENV 45.73082  3.85283195 -5.093371615  1.16967118 -0.08985538
#> LM-02    ENV 34.64462 -1.14575146 -0.881093222 -4.56547274  0.55159099
#> LM-03    ENV 53.83493  5.34625518  4.265275487 -0.14143931 -0.11714533
#> SR-02    ENV 14.95128 -2.58678337  0.660309540  0.89096920 -3.25055305
#> SR-03    ENV 11.15328 -3.17043379  0.082842050  0.68668051  0.15048221
#>                 PC5
#> 102.18  -0.04364115
#> 104.22   0.95312506
#> 121.31  -1.30661916
#> 141.28  -0.25996142
#> 157.26  -0.59719268
#> 163.9    0.18563390
#> 221.19  -0.57504816
#> 233.11   0.65754266
#> 235.6   -0.40273415
#> 241.2    0.07555258
#> 255.7   -0.46344763
#> 314.12   0.54406154
#> 317.6    0.39627052
#> 319.20   0.29657050
#> 320.16   2.29506737
#> 342.15  -0.10776433
#> 346.2   -0.12738693
#> 351.26   0.30191335
#> 364.21  -0.95811256
#> 402.7   -0.28473777
#> 405.2   -0.34397623
#> 406.12  -0.49796296
#> 427.7    1.00677993
#> 450.3   -0.34325251
#> 506.2    0.87807441
#> Canchan  0.49381313
#> Desiree -0.86767477
#> Unica   -0.90489253
#> Ayac     1.67177210
#> Hyo-02   0.01540152
#> LM-02    0.52350416
#> LM-03   -0.40285728
#> SR-02    1.37283488
#> SR-03   -3.18065538

# G*E matrix (deviations from mean)
array(model$genXenv, dim(model$genXenv), dimnames(model$genXenv))
#>          ENV
#> GEN              Ayac      Hyo-02       LM-02       LM-03        SR-02
#>   102.18    5.5726162 -12.4918224   1.7425251  -2.7070438   2.91734869
#>   104.22   -2.8712076   7.1684102   3.9336218  -4.0358373   0.47881580
#>   121.31    0.3255230  -3.8666836   4.3182811  10.4366135 -11.88343843
#>   141.28   -0.9451837   5.6454825  -9.7806639  14.6463104  -4.80337115
#>   157.26  -10.3149711 -10.6241677   4.2336365  16.8683612   2.71710210
#>   163.9     3.0874931  -6.9416721   3.4963790 -12.5533271   7.01688164
#>   221.19   -0.6041752  -6.0090018   4.0648518  -2.6974743   1.27671246
#>   233.11    2.5837535   6.8277609  -3.4440645  -4.4985717   0.19989490
#>   235.6    -1.7541523  19.8225025  -2.2394463  -5.6643239  -8.11400542
#>   241.2     1.0710975  -5.3831118   5.4253097  -3.2588271   0.46433086
#>   255.7     2.4443155   1.3860497  -1.8857757 -12.9626594   4.31373929
#>   314.12   -3.8812099   6.2098482   2.3577759   5.9071782  -3.92419060
#>   317.6    -1.7450319   3.0388540   3.0448064   5.5211634  -4.79271565
#>   319.20   -6.0155949   2.8477540  -9.7697504  24.8850017  -1.82949467
#>   320.16   10.9481796 -10.2982108   4.9608280  -6.2233088   2.99984918
#>   342.15    0.8508002  -0.3338618  -2.4575390 -10.3783871   7.29753151
#>   346.2     4.7000495  -6.2178087  -2.2612391 -14.9700672   9.90123888
#>   351.26    2.6002030  -0.9918665 -10.8315931  12.7429121  -0.02713985
#>   364.21   -0.4533734   3.2864208  -0.1335527  -0.1592533  -4.82292664
#>   402.7    -1.2134573  -0.0387229  -0.2179557  -0.8774011   1.08032472
#>   405.2     6.6477681  -8.3071271  -0.6159895  -8.8927189   3.52179705
#>   406.12   -6.1296667  12.0703469   1.1195092  -2.2601009  -3.13776595
#>   427.7    -3.1340922   4.3967072   4.2792028  -1.0194744   0.76266844
#>   450.3    -0.5047010  -1.0720791  -3.2821761  12.8806007  -5.04562407
#>   506.2    -1.2991912  -1.5682154   8.3142802  -3.1819279   0.60021498
#>   Canchan   1.2929442   5.7152780  -9.3713622   9.0803035  -1.65332869
#>   Desiree   9.5767845 -22.3280421   0.2396387 -11.8935722   9.62433886
#>   Unica   -10.8355195  18.0569790   4.7604622  -4.7341684  -5.13878822
#>          ENV
#> GEN             SR-03
#>   102.18    4.9663762
#>   104.22   -4.6738028
#>   121.31    0.6697043
#>   141.28   -4.7625741
#>   157.26   -2.8799609
#>   163.9     5.8942454
#>   221.19    3.9690870
#>   233.11   -1.6687730
#>   235.6    -2.0505746
#>   241.2     1.6812008
#>   255.7     6.7043306
#>   314.12   -6.6694018
#>   317.6    -5.0670763
#>   319.20  -10.1179157
#>   320.16   -2.3873373
#>   342.15    5.0214562
#>   346.2     8.8478267
#>   351.26   -3.4925156
#>   364.21    2.2826853
#>   402.7     1.2672123
#>   405.2     7.6462704
#>   406.12   -1.6623226
#>   427.7    -5.2850119
#>   450.3    -2.9760204
#>   506.2    -2.8651608
#>   Canchan  -5.0638348
#>   Desiree  14.7808522
#>   Unica    -2.1089651

# With default n (N') and default ssi.method (farshadfar)
ZA.AMMI(model)
#>                 Za SSI rZa rY    means
#> 102.18  0.15752787  41  18 23 26.31947
#> 104.22  0.08552245  20   7 13 31.28887
#> 121.31  0.13457796  26  11 15 30.10174
#> 141.28  0.20424009  23  22  1 39.75624
#> 157.26  0.20593889  28  23  5 36.95181
#> 163.9   0.16161024  46  19 27 21.41747
#> 221.19  0.08723440  34   8 26 22.98480
#> 233.11  0.06559491  21   4 17 28.66655
#> 235.6   0.20950908  29  25  4 38.63477
#> 241.2   0.08160010  28   6 22 26.34039
#> 255.7   0.16694984  34  20 14 30.58975
#> 314.12  0.12243347  28  10 18 28.17335
#> 317.6   0.08723605  18   9  9 35.32583
#> 319.20  0.30778801  30  27  3 38.75767
#> 320.16  0.14393358  35  14 21 26.34808
#> 342.15  0.13891478  37  13 24 26.01336
#> 346.2   0.20627243  49  24 25 23.84175
#> 351.26  0.17809076  29  21  8 36.11581
#> 364.21  0.03723882  12   2 10 34.05974
#> 402.7   0.01243185  20   1 19 27.47748
#> 405.2   0.15425031  33  17 16 28.98663
#> 406.12  0.13595705  24  12 12 32.68323
#> 427.7   0.07364374  12   5  7 36.19020
#> 450.3   0.14895835  22  16  6 36.19602
#> 506.2   0.06332050  14   3 11 33.26623
#> Canchan 0.14710608  35  15 20 27.00126
#> Desiree 0.32787182  56  28 28 16.15569
#> Unica   0.21646330  28  26  2 39.10400

# With n = 4 and default ssi.method (farshadfar)
ZA.AMMI(model, n = 4)
#>                 Za SSI rZa rY    means
#> 102.18  0.16239946  41  18 23 26.31947
#> 104.22  0.08993636  21   8 13 31.28887
#> 121.31  0.15679216  30  15 15 30.10174
#> 141.28  0.20676466  23  22  1 39.75624
#> 157.26  0.22558350  31  26  5 36.95181
#> 163.9   0.16668221  46  19 27 21.41747
#> 221.19  0.08837906  33   7 26 22.98480
#> 233.11  0.06788066  21   4 17 28.66655
#> 235.6   0.21970557  28  24  4 38.63477
#> 241.2   0.08459913  28   6 22 26.34039
#> 255.7   0.17014926  34  20 14 30.58975
#> 314.12  0.12303192  28  10 18 28.17335
#> 317.6   0.09305134  18   9  9 35.32583
#> 319.20  0.31897363  30  27  3 38.75767
#> 320.16  0.15713705  37  16 21 26.34808
#> 342.15  0.15011080  37  13 24 26.01336
#> 346.2   0.21536559  48  23 25 23.84175
#> 351.26  0.17844223  29  21  8 36.11581
#> 364.21  0.04502719  12   2 10 34.05974
#> 402.7   0.01603874  20   1 19 27.47748
#> 405.2   0.15936424  33  17 16 28.98663
#> 406.12  0.13981485  23  11 12 32.68323
#> 427.7   0.07895023  12   5  7 36.19020
#> 450.3   0.15508247  20  14  6 36.19602
#> 506.2   0.06378622  14   3 11 33.26623
#> Canchan 0.14787755  32  12 20 27.00126
#> Desiree 0.32833640  56  28 28 16.15569
#> Unica   0.22289692  27  25  2 39.10400

# With default n (N') and ssi.method = "rao"
ZA.AMMI(model, ssi.method = "rao")
#>                 Za       SSI rZa rY    means
#> 102.18  0.15752787 1.4309653  18 23 26.31947
#> 104.22  0.08552245 2.0752658   7 13 31.28887
#> 121.31  0.13457796 1.6519700  11 15 30.10174
#> 141.28  0.20424009 1.7380721  22  1 39.75624
#> 157.26  0.20593889 1.6429878  23  5 36.95181
#> 163.9   0.16161024 1.2566633  19 27 21.41747
#> 221.19  0.08723440 1.7838011   8 26 22.98480
#> 233.11  0.06559491 2.3102920   4 17 28.66655
#> 235.6   0.20950908 1.6903953  25  4 38.63477
#> 241.2   0.08160010 1.9646329   6 22 26.34039
#> 255.7   0.16694984 1.5378736  20 14 30.58975
#> 314.12  0.12243347 1.6556010  10 18 28.17335
#> 317.6   0.08723605 2.1861684   9  9 35.32583
#> 319.20  0.30778801 1.5568815  27  3 38.75767
#> 320.16  0.14393358 1.4859985  14 21 26.34808
#> 342.15  0.13891478 1.4977340  13 24 26.01336
#> 346.2   0.20627243 1.2148178  24 25 23.84175
#> 351.26  0.17809076 1.6842433  21  8 36.11581
#> 364.21  0.03723882 3.5336141   2 10 34.05974
#> 402.7   0.01243185 8.1540882   1 19 27.47748
#> 405.2   0.15425031 1.5301007  17 16 28.98663
#> 406.12  0.13595705 1.7293399  12 12 32.68323
#> 427.7   0.07364374 2.4052596   5  7 36.19020
#> 450.3   0.14895835 1.7859494  16  6 36.19602
#> 506.2   0.06332050 2.5096775   3 11 33.26623
#> Canchan 0.14710608 1.4937760  15 20 27.00126
#> Desiree 0.32787182 0.8019725  28 28 16.15569
#> Unica   0.21646330 1.6918583  26  2 39.10400

# Changing the ratio of weights for Rao's SSI
ZA.AMMI(model, ssi.method = "rao", a = 0.43)
#>                 Za       SSI rZa rY    means
#> 102.18  0.15752787 1.1044675  18 23 26.31947
#> 104.22  0.08552245 1.4738739   7 13 31.28887
#> 121.31  0.13457796 1.2697937  11 15 30.10174
#> 141.28  0.20424009 1.4862483  22  1 39.75624
#> 157.26  0.20593889 1.3932413  23  5 36.95181
#> 163.9   0.16161024 0.9384129  19 27 21.41747
#> 221.19  0.08723440 1.1942113   8 26 22.98480
#> 233.11  0.06559491 1.5261989   4 17 28.66655
#> 235.6   0.20950908 1.4449047  25  4 38.63477
#> 241.2   0.08160010 1.3343333   6 22 26.34039
#> 255.7   0.16694984 1.2298019  20 14 30.58975
#> 314.12  0.12243347 1.2355156  10 18 28.17335
#> 317.6   0.08723605 1.5965898   9  9 35.32583
#> 319.20  0.30778801 1.3897778  27  3 38.75767
#> 320.16  0.14393358 1.1286635  14 21 26.34808
#> 342.15  0.13891478 1.1274889  13 24 26.01336
#> 346.2   0.20627243 0.9654752  24 25 23.84175
#> 351.26  0.17809076 1.3954439  21  8 36.11581
#> 364.21  0.03723882 2.1524610   2 10 34.05974
#> 402.7   0.01243185 4.0169322   1 19 27.47748
#> 405.2   0.15425031 1.1966653  17 16 28.98663
#> 406.12  0.13595705 1.3510402  12 12 32.68323
#> 427.7   0.07364374 1.7068634   5  7 36.19020
#> 450.3   0.14895835 1.4406683  16  6 36.19602
#> 506.2   0.06332050 1.6974207   3 11 33.26623
#> Canchan 0.14710608 1.1441472  15 20 27.00126
#> Desiree 0.32787182 0.6451047  28 28 16.15569
#> Unica   0.21646330 1.4542544  26  2 39.10400